HeroListScreen Refactoring
Take out HeroItem from HeroListScreen
Take out BackgroundTriangles from HeroListScreen
Created Padding.kt for card padding
Created UiUtils.kt for screen height and width
Created Sizes.kt for card sizes
Take out text from HeroListScreen
