+++ app/src/main/java/com/example/marvelapp/ui/components/BackgroundTriangles.kt
@@ -1,0 +1,32 @@
+package com.example.marvelapp.ui.components
+
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Path
+
+@Composable
+fun BackgroundTriangles() {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Color.Black)
+    ) {
+        Canvas(
+            modifier = Modifier
+                .fillMaxSize()
+        ) {
+            val path = Path().apply {
+                moveTo(0f, size.height)
+                lineTo(size.width, size.height)
+                lineTo(size.width, size.height / 3)
+                close()
+            }
+            drawPath(path, Color.Red)
+        }
+    }
+}
