+++ app/src/main/java/com/example/marvelapp/ui/utils/Paddings.kt
@@ -1,0 +1,6 @@
+package com.example.marvelapp.ui.utils
+import androidx.compose.ui.unit.dp
+
+object Paddings {
+    val CardPadding = 16.dp
+}
