+++ app/src/main/java/com/example/marvelapp/ui/utils/Sizes.kt
@@ -1,0 +1,6 @@
+package com.example.marvelapp.ui.utils
+
+object Sizes {
+    val CardWidthFactor = 0.85f
+    val CardHeightFactor = 0.7f
+}
