+++ app/src/main/java/com/example/marvelapp/ui/utils/UiUtils.kt
@@ -1,0 +1,18 @@
+package com.example.marvelapp.ui.utils
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun screenWidthDp(): Dp {
+    val configuration = LocalConfiguration.current
+    return configuration.screenWidthDp.dp
+}
+
+@Composable
+fun screenHeightDp(): Dp {
+    val configuration = LocalConfiguration.current
+    return configuration.screenHeightDp.dp
+}
